\name{summaryStatistics}
\alias{summaryStatistics}
\title{ Basic function. }
\description{ 
   No description..
}

\usage{
-
}

\arguments{
  \item{x}{ vector of class numeric.}
}

\details{Basic function.
}

\value{
  \item{res}{ Result. }
}

\author{ Jiri Kalina \cr
         \email{kalina@mail.muni.cz}}

\examples{Try it yourself.
}
\keyword{  }
